requirements.txt
Django>=2.0,<3.0
djangorestframework>3.0
django-filter
django-cors-headers
setuptools>0
psycopg2>=0,<3.0
pytest
pytest-django
factory_boy
